+++ Game_alpha/cursor_t.cpp
@@ -13,7 +13,7 @@ cursor_t::cursor_t(std::string fileName, int _height, int _width, sf::RenderWind
 	texture.loadFromFile("img/cursor_aim.png");
 	sprite.setTexture(texture);
 	sprite.setTextureRect(sf::IntRect(0, 0, width, height));
-	sprite.setOrigin(width / 2, height / 2);
+	sprite.setOrigin(static_cast<float>(width) / 2, static_cast<float>(height) / 2);
 }
