+++ Game_alpha/ground_t.h
@@ -19,7 +19,7 @@ class ground_t
 
 public:
 	ground_t();
-	ground_t(sf::Texture *_texture,int _posX, int _posY, int _x, int _y, int _id);
+	ground_t(sf::Texture *_texture,float _posX, float _posY, int _x, int _y, int _id);
 	~ground_t();
